+++ app/Http/Controllers/Admin/PriorityCrudController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Requests\PriorityRequest;
+use App\Http\Requests\Admin\PriorityRequest;
 use Backpack\CRUD\app\Http\Controllers\CrudController;
 use Backpack\CRUD\app\Library\CrudPanel\CrudPanelFacade as CRUD;
