+++ app/Http/Requests/Admin/PriorityRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Admin;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UserProjectRequest extends FormRequest
+class PriorityRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -13,7 +13,8 @@ class UserProjectRequest extends FormRequest
      */
     public function authorize()
     {
-        return false;
+        // only allow updates if the user is logged in
+        return backpack_auth()->check();
     }
 
     /**
@@ -24,19 +25,7 @@ class UserProjectRequest extends FormRequest
     public function rules()
     {
         return [
-            //
-        ];
-    }
-
-    /**
-     * Get the validation attributes that apply to the request.
-     *
-     * @return array
-     */
-    public function attributes()
-    {
-        return [
-            //
+            'name' => 'required|unique:App\Models\Priority,name'
         ];
     }
 
@@ -48,7 +37,8 @@ class UserProjectRequest extends FormRequest
     public function messages()
     {
         return [
-            //
+            'name.required' => 'Навание обязательно к заполнению',
+            'name.unique' => 'Прироритет уже есть в БД'
         ];
     }
 }
